requirements.txt
# Universal Organization RAG System Dependencies
# Cross-platform compatible (M1/M2 Mac, Intel, Linux)

# CRITICAL: NumPy 1.x for compatibility with ML libraries  
numpy<2.0,>=1.24.0

# PyTorch - Cross-platform compatible versions
torch==2.2.2
torchvision==0.17.2

# Transformers ecosystem
transformers==4.38.0
sentence-transformers==2.6.1
huggingface-hub==0.20.3

# Vector database - Updated for MCP compatibility
chromadb>=1.0.15

# Configuration management
PyYAML>=6.0

# Text processing
tiktoken==0.6.0
Markdown==3.6

# System utilities
psutil==6.0.0
tqdm==4.66.4

# Core data science
pandas>=2.0.0

# HTTP requests
requests>=2.31.0

# Testing
pytest>=7.0.0

# MCP Integration for Claude Desktop
chroma-mcp>=0.2.5

# Optional: For advanced features
# jupyter>=1.0.0          # Notebook interface
# streamlit>=1.28.0       # Web UI
# gradio>=4.0.0          # Quick demos
